Update
Now the Animation is a GameObject, and now the objects need the Screen instead of the Game in the constructor. PaintIndex was added to objects to create "layers" in the screen rendering, now it doesn't matter in which position you add the object, if you edit this variable it will be painted before or after other objects.