Screen now has RenderOptions
Now the screens have a RenderOptions which you can define in the constructor, in case the screen is RenderOption.WHEN_NEED the screen will only be rendered when it needs it (variable need_render) in addition the objects will also be rendered only when necessary, using the same variable but this time inside the object. An empty method (renderTick) has also been added which can be modified to add logic in case it is necessary to render the object or screen as necessary. If you set the screen's RenderOption to PER_FRAME everything will render at the estimated FPS all the time.